LPS-123793 Moves handler functions to utility files to be reused and exposes the useful methods from Fields and SettingsContext of LayoutProvider
