+++ modules/apps/dynamic-data-mapping/dynamic-data-mapping-form-builder/src/main/resources/META-INF/resources/js/components/LayoutProvider/handlers/fieldDeletedHandler.es.js
@@ -12,82 +12,10 @@
  * details.
  */
 
-import {FormSupport, PagesVisitor} from 'dynamic-data-mapping-form-renderer';
+import {FormSupport} from 'dynamic-data-mapping-form-renderer';
 
-import {FIELD_TYPE_FIELDSET} from '../../../util/constants.es';
+import {removeField} from '../../../util/fieldSupport.es';
 import RulesSupport from '../../RuleBuilder/RulesSupport.es';
-import {updateField} from '../util/settingsContext.es';
-
-export const removeField = (
-	props,
-	pages,
-	fieldName,
-	removeEmptyRows = true
-) => {
-	const visitor = new PagesVisitor(pages);
-
-	const filter = (fields) =>
-		fields
-			.filter((field) => field.fieldName !== fieldName)
-			.map((field) => {
-				const nestedFields = field.nestedFields
-					? filter(field.nestedFields)
-					: [];
-
-				field = updateField(props, field, 'nestedFields', nestedFields);
-
-				if (field.type !== FIELD_TYPE_FIELDSET) {
-					return {
-						...field,
-						nestedFields,
-					};
-				}
-
-				let rows = [];
-
-				if (field.rows) {
-					const visitor = new PagesVisitor([
-						{
-							rows:
-								typeof field.rows === 'string'
-									? JSON.parse(field.rows)
-									: field.rows || [],
-						},
-					]);
-
-					const pages = visitor.mapColumns((column) => ({
-						...column,
-						fields: column.fields.filter(
-							(nestedFieldName) => fieldName !== nestedFieldName
-						),
-					}));
-
-					rows = removeEmptyRows
-						? FormSupport.removeEmptyRows(pages, 0)
-						: pages[0].rows;
-
-					field = updateField(props, field, 'rows', rows);
-				}
-
-				return {
-					...field,
-					nestedFields,
-					rows,
-				};
-			})
-			.filter(({nestedFields = [], type}) => {
-				if (type === FIELD_TYPE_FIELDSET && !nestedFields.length) {
-					return false;
-				}
-
-				return true;
-			});
-
-	return visitor.mapColumns((column) => ({
-		...column,
-		fields: filter(column.fields),
-	}));
-};
 
 export const handleFieldDeleted = (
 	props,
