+++ modules/apps/dynamic-data-mapping/dynamic-data-mapping-form-builder/src/main/resources/META-INF/resources/js/components/LayoutProvider/handlers/fieldDuplicatedHandler.es.js
@@ -12,181 +12,10 @@
  * details.
  */
 
-import {
-	FormSupport,
-	PagesVisitor,
-	generateInstanceId,
-} from 'dynamic-data-mapping-form-renderer';
+import {FormSupport, PagesVisitor} from 'dynamic-data-mapping-form-renderer';
 
-import {
-	getDefaultFieldName,
-	localizeField,
-} from '../../../util/fieldSupport.es';
-import {sub} from '../../../util/strings.es';
-import {getFieldLocalizedValue} from '../util/fields.es';
-import {
-	getSettingsContextProperty,
-	updateField,
-	updateFieldLabel,
-	updateSettingsContextInstanceId,
-	updateSettingsContextProperty,
-} from '../util/settingsContext.es';
-
-export const getLabel = (
-	originalField,
-	defaultLanguageId,
-	editingLanguageId
-) => {
-	const labelFieldLocalizedValue = getFieldLocalizedValue(
-		originalField.settingsContext.pages,
-		'label',
-		editingLanguageId
-	);
-
-	if (!labelFieldLocalizedValue) {
-		return;
-	}
-
-	return sub(Liferay.Language.get('copy-of-x'), [labelFieldLocalizedValue]);
-};
-
-export const getValidation = (originalField) => {
-	const validation = getSettingsContextProperty(
-		originalField.settingsContext,
-		'validation'
-	);
-
-	return validation;
-};
-
-export const createDuplicatedField = (originalField, props, blacklist = []) => {
-	const {
-		availableLanguageIds,
-		defaultLanguageId,
-		editingLanguageId,
-		fieldNameGenerator,
-		generateFieldNameUsingFieldLabel,
-	} = props;
-	const newFieldName = fieldNameGenerator(
-		getDefaultFieldName(),
-		null,
-		blacklist
-	);
-
-	let duplicatedField = updateField(
-		props,
-		originalField,
-		'name',
-		newFieldName
-	);
-
-	duplicatedField = updateField(
-		props,
-		duplicatedField,
-		'fieldReference',
-		newFieldName
-	);
-
-	duplicatedField.instanceId = generateInstanceId(8);
-
-	availableLanguageIds.forEach((availableLanguageId) => {
-		const label = getLabel(
-			originalField,
-			defaultLanguageId,
-			availableLanguageId
-		);
-
-		if (label) {
-			duplicatedField = updateFieldLabel(
-				defaultLanguageId,
-				availableLanguageId,
-				fieldNameGenerator,
-				duplicatedField,
-				generateFieldNameUsingFieldLabel,
-				label
-			);
-		}
-	});
-
-	if (duplicatedField.nestedFields?.length > 0) {
-		duplicatedField.nestedFields = duplicatedField.nestedFields.map(
-			(field) => {
-				const newDuplicatedNestedField = createDuplicatedField(
-					field,
-					props,
-					blacklist
-				);
-
-				blacklist.push(newDuplicatedNestedField.fieldName);
-
-				let {rows = []} = duplicatedField;
-
-				if (typeof rows === 'string') {
-					rows = JSON.parse(rows);
-				}
-
-				const visitor = new PagesVisitor([
-					{
-						rows,
-					},
-				]);
-
-				const layout = visitor.mapColumns((column) => {
-					return {
-						...column,
-						fields: column.fields.map((fieldName) => {
-							if (fieldName === field.fieldName) {
-								return newDuplicatedNestedField.fieldName;
-							}
-
-							return fieldName;
-						}),
-					};
-				});
-
-				duplicatedField.rows = layout[0].rows;
-
-				return newDuplicatedNestedField;
-			}
-		);
-
-		duplicatedField.settingsContext = updateSettingsContextProperty(
-			defaultLanguageId,
-			props.editingLanguageId,
-			duplicatedField.settingsContext,
-			'rows',
-			duplicatedField.rows
-		);
-
-		duplicatedField.ddmStructureLayoutId = '';
-
-		duplicatedField.settingsContext = updateSettingsContextProperty(
-			defaultLanguageId,
-			props.editingLanguageId,
-			duplicatedField.settingsContext,
-			'ddmStructureLayoutId',
-			duplicatedField.ddmStructureLayoutId
-		);
-	}
-
-	const settingsContext = updateSettingsContextInstanceId(duplicatedField);
-
-	const settingsVisitor = new PagesVisitor(settingsContext.pages);
-
-	duplicatedField.settingsContext = {
-		...settingsContext,
-		pages: settingsVisitor.mapFields((field) =>
-			localizeField(field, defaultLanguageId, editingLanguageId)
-		),
-	};
-
-	return updateField(
-		props,
-		duplicatedField,
-		'validation',
-		getValidation(duplicatedField)
-	);
-};
+import {createDuplicatedField} from '../util/fields.es';
+import {updateField} from '../util/settingsContext.es';
 
 export const duplicateField = (
 	activePage,
