+++ modules/apps/dynamic-data-mapping/dynamic-data-mapping-form-builder/src/main/resources/META-INF/resources/js/components/LayoutProvider/util/fields.es.js
@@ -15,10 +15,19 @@
 import {
 	FormSupport,
 	PagesVisitor,
+	generateInstanceId,
 	normalizeFieldName,
 } from 'dynamic-data-mapping-form-renderer';
 
-import {getDefaultFieldName} from '../../../util/fieldSupport.es';
+import {getDefaultFieldName, localizeField} from '../../../util/fieldSupport.es';
+import {sub} from '../../../util/strings.es';
+import {
+	getSettingsContextProperty,
+	updateField,
+	updateFieldLabel,
+	updateSettingsContextInstanceId,
+	updateSettingsContextProperty,
+} from '../util/settingsContext.es';
 
 export const generateFieldName = (
 	pages,
@@ -114,6 +123,24 @@ export const getFieldLocalizedValue = (pages, fieldName, locale) => {
 	return fieldLocalizedValue[locale];
 };
 
+export const getLabel = (
+	originalField,
+	defaultLanguageId,
+	editingLanguageId
+) => {
+	const labelFieldLocalizedValue = getFieldLocalizedValue(
+		originalField.settingsContext.pages,
+		'label',
+		editingLanguageId
+	);
+
+	if (!labelFieldLocalizedValue) {
+		return;
+	}
+
+	return sub(Liferay.Language.get('copy-of-x'), [labelFieldLocalizedValue]);
+};
+
 export const updateFieldValidationProperty = (
 	pages,
 	fieldName,
@@ -152,4 +179,164 @@ export const updateFieldValidationProperty = (
 
 		return field;
 	});
+};
+
+export const getValidation = (originalField) => {
+	const validation = getSettingsContextProperty(
+		originalField.settingsContext,
+		'validation'
+	);
+
+	return validation;
+};
+
+export const createDuplicatedField = (originalField, props, blacklist = []) => {
+	const {
+		availableLanguageIds,
+		defaultLanguageId,
+		editingLanguageId,
+		fieldNameGenerator,
+		generateFieldNameUsingFieldLabel,
+	} = props;
+	const newFieldName = fieldNameGenerator(
+		getDefaultFieldName(),
+		null,
+		blacklist
+	);
+
+	let duplicatedField = updateField(
+		props,
+		originalField,
+		'name',
+		newFieldName
+	);
+
+	duplicatedField = updateField(
+		props,
+		duplicatedField,
+		'fieldReference',
+		newFieldName
+	);
+
+	duplicatedField.instanceId = generateInstanceId(8);
+
+	availableLanguageIds.forEach((availableLanguageId) => {
+		const label = getLabel(
+			originalField,
+			defaultLanguageId,
+			availableLanguageId
+		);
+
+		if (label) {
+			duplicatedField = updateFieldLabel(
+				defaultLanguageId,
+				availableLanguageId,
+				fieldNameGenerator,
+				duplicatedField,
+				generateFieldNameUsingFieldLabel,
+				label
+			);
+		}
+	});
+
+	if (duplicatedField.nestedFields?.length > 0) {
+		duplicatedField.nestedFields = duplicatedField.nestedFields.map(
+			(field) => {
+				const newDuplicatedNestedField = createDuplicatedField(
+					field,
+					props,
+					blacklist
+				);
+
+				blacklist.push(newDuplicatedNestedField.fieldName);
+
+				let {rows = []} = duplicatedField;
+
+				if (typeof rows === 'string') {
+					rows = JSON.parse(rows);
+				}
+
+				const visitor = new PagesVisitor([
+					{
+						rows,
+					},
+				]);
+
+				const layout = visitor.mapColumns((column) => {
+					return {
+						...column,
+						fields: column.fields.map((fieldName) => {
+							if (fieldName === field.fieldName) {
+								return newDuplicatedNestedField.fieldName;
+							}
+
+							return fieldName;
+						}),
+					};
+				});
+
+				duplicatedField.rows = layout[0].rows;
+
+				return newDuplicatedNestedField;
+			}
+		);
+
+		duplicatedField.settingsContext = updateSettingsContextProperty(
+			defaultLanguageId,
+			props.editingLanguageId,
+			duplicatedField.settingsContext,
+			'rows',
+			duplicatedField.rows
+		);
+
+		duplicatedField.ddmStructureLayoutId = '';
+
+		duplicatedField.settingsContext = updateSettingsContextProperty(
+			defaultLanguageId,
+			props.editingLanguageId,
+			duplicatedField.settingsContext,
+			'ddmStructureLayoutId',
+			duplicatedField.ddmStructureLayoutId
+		);
+	}
+
+	const settingsContext = updateSettingsContextInstanceId(duplicatedField);
+
+	const settingsVisitor = new PagesVisitor(settingsContext.pages);
+
+	duplicatedField.settingsContext = {
+		...settingsContext,
+		pages: settingsVisitor.mapFields((field) =>
+			localizeField(field, defaultLanguageId, editingLanguageId)
+		),
+	};
+
+	return updateField(
+		props,
+		duplicatedField,
+		'validation',
+		getValidation(duplicatedField)
+	);
+};
+
+export const findInvalidFieldReference = (focusedField, pages, value) => {
+	let hasInvalidFieldReference = false;
+
+	const visitor = new PagesVisitor(pages);
+
+	visitor.mapFields((field) => {
+		const fieldReference = getSettingsContextProperty(
+			field.settingsContext,
+			'fieldReference'
+		);
+
+		if (
+			focusedField.fieldName !== field.fieldName &&
+			fieldReference === value
+		) {
+			hasInvalidFieldReference = true;
+		}
+	});
+
+	return hasInvalidFieldReference;
 };
