+++ modules/apps/dynamic-data-mapping/dynamic-data-mapping-form-builder/src/main/resources/META-INF/resources/js/index.es.js
@@ -19,4 +19,9 @@ import * as FieldSupport from './util/fieldSupport.es';
 export default FormBuilderWithLayoutProvider;
 export {Token} from '../js/expressions/Token.es';
 export {Tokenizer} from '../js/expressions/Tokenizer.es';
+export * as SettingsContext from '../js/components/LayoutProvider/util/settingsContext.es';
+export * as RulesUtil from '../js/components/LayoutProvider/util/rules.es';
+export * as FieldUtil from '../js/components/LayoutProvider/util/fields.es';
+export * as FieldSetUtil from '../js/components/LayoutProvider/util/fieldset.es';
 export {FieldSupport, RulesSupport};
+export {default as Sidebar} from './components/Sidebar/Sidebar.es';
